requirements.txt
openai>=1.12.0
whisper>=1.0.0
tiktoken>=0.5.0
pydub>=0.25.1
requests>=2.31.0
pytest>=7.4.0
python-dotenv>=1.0.0
pyyaml>=6.0.1
structlog>=24.1.0
paho-mqtt>=2.1.0
feedparser>=6.0.10
fastapi>=0.110.0
uvicorn>=0.27.0
python-multipart>=0.0.9
aiohttp>=3.9.1 